fix(plugin-multi-tenant): issue #10740 - "The following field is invalid: Assigned Tenant" (#10764)
### What?
Tenant ID is a `number` but the `beforeChange` hook shows it as a `string`.
Getting tenant ID from cookie does not work properly in PG

### Why?
A `ValidationError` is throwing when reading a pg numeric id from the cookie.

### How?
Adjust the id based on the idType on the collection.

Fixes #10740

---------

+++ packages/plugin-multi-tenant/src/utilities/getTenantFromCookie.ts
@@ -13,5 +13,5 @@ export function getTenantFromCookie(
 ): null | number | string {
   const cookies = parseCookies(headers)
   const selectedTenant = cookies.get('payload-tenant') || null
-  return selectedTenant ? (idType === 'number' ? parseInt(selectedTenant) : selectedTenant) : null
+  return selectedTenant ? (idType === 'number' ? parseFloat(selectedTenant) : selectedTenant) : null
 }
